Avoid binding several types at once to the same scope, preparing Coq PR#7762.
On one side, this is not documented. On another side, this is
misleading as shown in Coq issue #7699.  On a third side, the case of
several types binding to the same scope looks unusual. Finally, this
may leave a syntax open for a generalization to "Bind Scope foo with
type_pattern" rather than having to add parentheses as in "Bind Scope
foo with (type_pattern) ... (type_pattern)".