[omega] Remove use of undocumented from `omega with *`
This is to help https://github.com/coq/coq/pull/11288 , in most cases
we just change `omega with *` to `zify; omega`. In a few cases we had
to call `lia`, but this should be cleaned up when upstream merges a
few more micromega / omega PRs.

Performance impact should be minimal.